Move remaining shared utils to utils
