+++ projects/js-packages/charts/src/components/leaderboard-chart/index.tsx
@@ -1,4 +1,4 @@
 export { default as LeaderboardChart } from './leaderboard-chart';
 export type { LeaderboardChartProps, LeaderboardEntry } from './leaderboard-chart';
-export { formatMetricValue } from '../shared/format-metric-value';
-export type { MetricValueType } from '../shared/format-metric-value';
+export { formatMetricValue } from '../../utils/format-metric-value';
+export type { MetricValueType } from '../../utils/format-metric-value';
