+++ projects/js-packages/charts/src/components/leaderboard-chart/leaderboard-chart.tsx
@@ -8,7 +8,7 @@ import { Fragment } from '@wordpress/element';
 import clsx from 'clsx';
 import { type FC } from 'react';
 import { useGlobalChartTheme } from '../../hooks';
-import { formatMetricValue } from '../shared/format-metric-value';
+import { formatMetricValue } from '../../utils/format-metric-value';
 import styles from './leaderboard-chart.module.scss';
 
 /**
