+++ projects/js-packages/charts/src/utils/format-metric-value.ts
